refactor(cli): extract review-card spawn into work_commands_review.rs (card c0bd865c phase 5 / a3ede6de)
## What this PR lands

Phase 5 of the 5-way split. The review-card spawn — both the manual `airc work review <parent>` CLI path (ad7e100b Sub-B) and the auto-spawn that fires on `state review` (ad7e100b Sub-C) — moves to a sibling module:

  Extracted to work_commands_review.rs:
    - run_review — CLI handler for `airc work review`
    - format_review_title — `review: <parent>` (truncate at 80)
    - auto_spawn_review_card — fires from open_pr_and_link on state→Review

Both paths produce structurally-identical review cards via the shared title format, so observers that filter on `title.starts_with("review:")` pick up either.

## File sizes

  work_commands.rs:        1670 → 1517 LOC (-153 this PR)
  work_commands_review.rs:    0 →  155 LOC (new)

## Cumulative across all 5 phases (when they all land)

  Original:                          1823 LOC
  - phase 1 (#1042, render):         -238
  - phase 2 (#1046, lifecycle):      -374
  - phase 3 (#1047, git):            -189
  - phase 5 (this PR, review):       -153
  - phase 4 still TODO (c7bdf2df gh): -~250 expected
  - Final after all 5:               ~620 LOC of pure dispatchers + shared types

That's the goal: a sub-thousand-line work_commands.rs where consumer-embedders (continuum / hermes / openclaw / codex when planning integrations) can scan the public surface in one pass.

## Cross-module wiring

  - open_pr_and_link (in work_commands.rs, will move to gh in phase 4) → auto_spawn_review_card (this PR, pub(crate))
  - tests in work_commands.rs → format_review_title (this PR, called via crate::work_commands_review::)

`run_review` is re-exported in work_commands.rs as `pub use` so main.rs's existing `work_commands::run_review(...)` dispatch is unchanged.

## Tests + lints

  cargo test work_commands:: → 19 pass (3 review-title tests included: review_title_preserves_short_parent_title_verbatim / counts_chars_not_bytes_for_unicode_parents / truncates_long_parent_with_ellipsis_marker)
  cargo clippy -p airc-cli --bin airc -- -D warnings → clean
  cargo fmt --check → clean

## Anchor

c0bd865c follow-up sequence. Phase 1 = render (#1042). Phase 2 = lifecycle (#1046). Phase 3 = git (#1047). Phase 5 = review (this PR). Phase 4 = gh (c7bdf2df) is the last extraction; once it lands the file is done.
